# backend/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: CERBERUS API
  description: API definition for the CERBERUS implant-management API.
  version: 2.0.0
# TODO Update with ACG stuff
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
  schemas:
...
    taskreq:
      properties:
        implantId:
          nullable: false
          type: "string"
        taskType:
          nullable: false
          type: "string"
        params:
          nullable: true
          type: "array"
          items:
            $ref: "#/components/schemas/paramvalue"
    taskdb:
      properties:
        _id:
          nullable: false
          type: "string"
        order:
          nullable: false
          type: "integer"
        implantId:
          nullable: false
          type: "string"
        taskType:
          nullable: false
          type: "string"
        params:
          nullable: true
          type: "array"
          items:
            $ref: "#/components/schemas/paramvalue"
        sent:
          nullable: false
          type: "boolean"
    tasktypedb:
      properties:
        _id:
          nullable: false
          type: "string"
        name:
          nullable: false
          type: "string"
        params:
          nullable: false
          type: "array"
          items:
            type: object
            properties:
              name: 
                nullable: false
                type: "string"
              type: 
                nullable: false
                type: "string"
    tasktypecreate:
      properties:
        name:
          nullable: false
          type: "string"
        params:
          nullable: false
          type: "array"
          items:
            type: object
            properties:
              name: 
                nullable: false
                type: "string"
              type: 
                nullable: false
                type: "string"
    paramvalue:
      properties:
        name:
          nullable: false
          type: "string"
        value:
          nullable: false
          type: "string"
paths:
...
  /tasks/{implantId}:
    get:
      security:
        - jwt: [Read]
      summary: Retrieves tasks for the given implant. Requires read access to the implant.
      tags:
        - Tasks
        - Implants
      parameters:
        - in: path
          name: implantId
          schema:
            type: string
          required: true
          description: Identifier of the implant to find tasks from.
      responses:
        "200": # status code
          description: Tasks.
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    description: Tasks for the implant
                    items:
                      $ref: "#/components/schemas/taskdb"
                  errors:
                    type: array
                    items:
                      type: string
        "403": # status code
          description: User not logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "500": # status code
          description: An unexpected server-side error
          content:
            application/json:
              schema:
                type: object
                properties:
                  implants:
                    type: array
                    description: This will be empty
                  errors:
                    type: array
                    items:
                      type: string
  /task-types:
    get:
      security:
        - jwt: [Read]
      summary: Retrieves the list of all task types
      tags:
        - Tasks
      responses:
        "200": # status code
          description: Task types
          content:
            application/json:
              schema:
                type: object
                properties:
                  taskTypes:
                    type: array
                    description: Task types
                    items:
                      $ref: "#/components/schemas/tasktypedb"
                  errors:
                    type: array
                    items:
                      type: string
        "403": # status code
          description: User not logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "500": # status code
          description: An unexpected server-side error
          content:
            application/json:
              schema:
                type: object
                properties:
                  implants:
                    type: array
                    description: This will be empty
                  errors:
                    type: array
                    items:
                      type: string
    post:
      security:
        - jwt: [Admin]
      summary: Creates a new task type
      tags:
        - Tasks
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                taskType:
                  $ref: "#/components/schemas/tasktypecreate"            
      responses:
        "200": # status code
          description: Successfully created
          content:
            application/json:
              schema:
                type: object
                properties:
                  taskType: 
                    type: object
                    $ref: "#/components/schemas/tasktypedb"
                  errors:
                    type: array
                    items:
                      type: string
        "403": # status code
          description: User not logged in or not an admin
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "500": # status code
          description: An unexpected server-side error
          content:
            application/json:
              schema:
                type: object
                properties:
                  taskType:
                    type: object
                    description: This will be empty
                  errors:
                    type: array
                    items:
                      type: string
  /task-types/{taskTypeId}:
    delete:
      security:
        - jwt: [Admin]
      summary: Deletes a task type. Requires admin access.
      tags:
        - Tasks
      parameters:
        - in: path
          name: taskTypeId
          schema:
            type: string
          required: true
          description: Identifier of the task type to delete.
      responses:
        "200": # status code
          description: Successfully deleted (or did not exist in the first place)
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "403": # status code
          description: User not logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "500": # status code
          description: An unexpected server-side error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
...
  /tasks: 
    post: 
      security:
        - jwt: [Operator]
      summary: Creates or edits a task. Requires operator access to the implant.
      tags:
        - Tasks
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/taskreq"
      responses:
        "200": # status code
          description: Successfully created/updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "400": # status code
          description: Improperly-formatted task
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "403": # status code
          description: User not logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "500": # status code
          description: An unexpected server-side error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
  /tasks/{taskId}: 
    delete: 
      security:
        - jwt: [Operator]
      summary: Deletes a task. Requires operator access to the implant.
      tags:
        - Tasks
      parameters:
        - in: path
          name: taskId
          schema:
            type: string
          required: true
          description: Identifier of the task to delete.
      responses:
        "200": # status code
          description: Successfully deleted (or did not exist in the first place)
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "400": # status code
          description: Task has already been sent to an implant
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "403": # status code
          description: User not logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "500": # status code
          description: An unexpected server-side error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
# USERS
...